Clean up and organize files
