+++ src/components/App/App.js
@@ -1,3 +1,4 @@
+import './App.css'
 import { useState } from 'react';
 
 function App() {
